v_gamma/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "robotics-presentation-suite"
version = "1.0.0"
description = "Industry-standard robotics simulation presentation pipeline"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Robotics Team", email = "robotics@company.com"}
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Robotics",
]
requires-python = ">=3.9"
dependencies = [
    "pybullet>=3.2.5",
    "numpy>=1.21.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "matplotlib>=3.5.0",
    "plotly>=5.0.0",
    "dash>=2.0.0",
    "opencv-python>=4.5.0",
    "pillow>=9.0.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
docs = [
    "sphinx>=6.0.0",
    "sphinx-rtd-theme>=1.2.0",
    "myst-parser>=2.0.0",
]

[project.scripts]
robotics-demo = "robotics_suite.cli:main"
robotics-server = "robotics_suite.server:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src/robotics_suite --cov-report=term-missing --cov-report=html --cov-report=xml --strict-markers --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests"
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
    "ignore::PendingDeprecationWarning"
]